+++ projects/progress-widget/src/app/progress-bar/works-in-progress/work-in-progress.ts
@@ -16,7 +16,7 @@ export class WorkInProgress{
             this.percentage = 20;
         }
         else{
-            this.percentage = Math.floor((this.writingProject.status/8)*100);
+            this.percentage = Math.floor((this.writingProject.status/9)*100);
         }
     }