+++ projects/progress-widget/src/app/progress-bar/works-in-progress/works-in-progress.component.ts
@@ -23,7 +23,7 @@ export class WorksInProgressComponent implements OnInit {
         appConfigService.getProgressBarHeading()
           .subscribe(
             data => {
-              console.log("Grabbing text for the heading");
+              console.log("Setting the text for the heading");
               this.heading = data;
             }
           )
@@ -34,7 +34,7 @@ export class WorksInProgressComponent implements OnInit {
         progressBarService.getWorksInProgress()
           .subscribe(
             data => {
-              console.log("Retrieved Works in Progress from the API");
+              console.log("Setting the list of the Works in Progress from the API");
               this.worksInProgress = data.map( 
                 item => new WorkInProgress(item)
               );