+++ changelog.md
@@ -4,6 +4,8 @@
 
 - Added new value to the `Staus` enum
 - Updated the progress widget heading to `h2` 
+- Updated the progress widget heading to use configurable text for the heading
+- Moved the Heading out of the main body of the widget
 
 ## 1.2.4