+++ changelog.md
@@ -1,8 +1,8 @@
 # SmBlogWidgets
 
 ## vNext
-
-- Added new value to the `Staus` enum
+- Updated the padding around the quote
+- Added new value to the `Status` enum
 - Updated the progress widget heading to `h2` 
 - Updated the progress widget heading to use configurable text for the heading
 - Moved the Heading out of the main body of the widget